refactor(metrics): extract route_summary module
